Fix two important problems
- Fix the problem of the maximum allowable number of characters in
a telegram message. Now the message will be sent if its number of
characters does not exceed 4096. Otherwise, the message will be sent
in parts so as not to cause a telegram error.

- Fix the problem of the maximum allowable number of inline buttons
in the message. Now if the user has more than 11 products, inline
buttons will appear with the ability to go to the next or previous
page of their products.

- Add new imports

- Now the entire logic of "viewing your products" takes place in the
FSM state "Product_review_form" for correct and quick switching
between product pages